dtype fix #2

+++ src/inference.py
@@ -72,17 +72,17 @@ def validate():
 
             a = fig.add_subplot(1,3,1)
             input_imx = input_image.detach().cpu().numpy().reshape(3, 224, 224)
-            plt.imshow(np.asarray(np.transpose(input_imx, (1,2,0)), dtype=int))
+            plt.imshow(np.asarray(np.transpose(input_imx, (1,2,0)), dtype=np.short))
             a.set_title('Input Image')
             
             a = fig.add_subplot(1,3,2)
             predicted_mx = predicted_mask.detach().cpu().numpy().reshape(3, 224, 224)
-            plt.imshow(np.asarray(np.transpose(predicted_mx, (1,2,0)), dtype=int))
+            plt.imshow(np.asarray(np.transpose(predicted_mx, (1,2,0)), dtype=np.short))
             a.set_title('Predicted Mask')
 
             a = fig.add_subplot(1,3,3)
             target_mx = target_mask.detach().cpu().numpy().reshape(3, 224, 224)
-            plt.imshow(np.asarray(np.transpose(target_mx, (1,2,0)), dtype=int))
+            plt.imshow(np.asarray(np.transpose(target_mx, (1,2,0)), dtype=np.short))
             a.set_title('Ground Truth')
 
             fig.savefig(os.path.join(OUTPUT_DIR, "prediction_{}_{}.png".format(batch_idx, idx)))
